+++ packages/appcd-plugin/CHANGELOG.md
@@ -11,6 +11,7 @@
  * feat(scheme): Added `detect()` method to schemes that resolves a list of plugins found. Also
    managed to clean up a bunch of code and tests in the process.
  * feat: Added symlink flag to plugin descriptor. Used to guess if plugin is a yarn link.
+ * feat: Added support for a `autoStart` flag in the plugin's `package.json`.
  * fix: Fixed bug where `status` would search an unsorted list of registered plugins which caused
    nondeterministic results. [(DAEMON-328)](https://jira.appcelerator.org/browse/DAEMON-328)
  * chore: Updated dependencies.